requirements.txt
# Isaac AI-Robot Brain Module Requirements

# Core Python dependencies
numpy>=1.21.0
scipy>=1.7.0
opencv-python>=4.5.0
Pillow>=8.3.0
pycuda>=2022.1

# ROS 2 Python interfaces
rclpy>=3.0.0
sensor_msgs
geometry_msgs
nav_msgs

# Isaac ROS packages
isaac_ros_visual_slam
isaac_ros_stereo_image_rectification

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=0.19.0

# Logging
structlog>=21.0.0

# USD and Omniverse integration
omni-client